call nextDouble() since method returns a double

+++ src/Console.java
@@ -7,7 +7,7 @@ public class Console {
 
         while (true) {
             System.out.print(prompt);
-            value = scanner.nextFloat();
+            value = scanner.nextDouble();
             if (value >= min && value <= max)
                 break;
             System.out.println("Enter a value between " + min + " and " + max);
